requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
sentence-transformers
faiss-cpu
redis
